Corrected data types for file descriptors.

+++ Blatt 8 - Datein und Speicherzugriffe/A13 mycp/mycp.c
@@ -1,20 +1,27 @@
 #include "mycp.h"
 
-size_t source;
-size_t target;
+int source;
+int target;
 
 void main( int argc, char* argv[] ) {
 
-  //initialize
+  //initialize?
 
   //main stuff:
 
   openSource( char *path[] );
+  // handle: file does not exist
   createTarget( char *path[] );
-  getSourceLength( size_t fileDescriptor );
+  // handle: file already exists;
+  // handle: file could not be created;
+  getSourceLength( int fileDescriptor );
+  // handle: generic error
   setTargetSize( size_t targetSize );
-  mapFileToMemory( size_t fileDescriptor );
-  copyFilesInMemory( size_t source, size_t target );
+  // handle: generic error
+  mapFileToMemory( int fileDescriptor );
+  // handle: generic error
+  copyFilesInMemory( int source, int target );
+  // handle: generic error
 
   //cleanup:
     //munmap both files
